feat(PI-4628): initial implementation of 'run package'

+++ internal/platformapi/client.oapi.go
@@ -6832,7 +6832,7 @@ type PostAppPlatformPackageAssessmentResponse struct {
 		} `json:"status"`
 
 		// Task An integer identifying the underlying analysis tasks associated with this assessment.
-		Task float32 `json:"task"`
+		Task float64 `json:"task"`
 
 		// TaskErrorCode The error code describing the failure if an assessment failed.
 		TaskErrorCode *string `json:"task_error_code"`
@@ -10062,7 +10062,7 @@ func ParsePostAppPlatformPackageAssessmentResponse(rsp *http.Response) (*PostApp
 			} `json:"status"`
 
 			// Task An integer identifying the underlying analysis tasks associated with this assessment.
-			Task float32 `json:"task"`
+			Task float64 `json:"task"`
 
 			// TaskErrorCode The error code describing the failure if an assessment failed.
 			TaskErrorCode *string `json:"task_error_code"`
